backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.98
botocore==1.34.98
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
daphne==4.0.0
Django==4.0
django-autoreload==0.1.1
django-cors-headers==4.3.0
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.13.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
load-dotenv==0.1.0
MarkupSafe==2.1.4
msgpack==1.0.7
mysqlclient==2.2.4
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyinotify==0.9.6
PyJWT==2.8.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
s3transfer==0.10.1
selenium==4.16.0
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websocket-server==0.6.4
wsproto==1.2.0
zope.interface==6.1
